Add basic data input (#7)
Adds the ability to transform a DataObject into a DataInput. This will add an ioSpecification to the parent if needed. Currently, creation, revert, redo, deletion, export, and import are supported.

DataInputs cannot yet be moved between different parents, connected to anything or be set as targets for DataObjects. That will be targeted in following PRs.